DataTables

+++ Routes/Clientes.py
@@ -1,6 +1,8 @@
-from flask import Blueprint,render_template,request,redirect,url_for,jsonify
+from flask import Blueprint,render_template,request,redirect,url_for,jsonify,flash
 from Models.Clientes_Models import clientes
 from Utils.db import db
+import re
+
 
 Clientes=Blueprint('Clientes',__name__)
 
@@ -8,7 +10,38 @@ Clientes=Blueprint('Clientes',__name__)
 def InicioCliente():
     
     Clientes = clientes.query.all()
-    return render_template('Clientes/Inicio.html',Clientes=Clientes)
+    clientes_array = []
+    for cliente in Clientes:
+        cliente_dict = {
+            'nombre': cliente.Nombre,
+            'cedula': cliente.Cedula,
+            'telefono': cliente.Telefono,
+            'direccion': cliente.Direccion,
+        }
+        clientes_array.append(cliente_dict)
+    return render_template('Clientes/Inicio.html', Clientes=jsonify(clientes_array))
+
+@Clientes.route('/dtCustomer', methods = ['GET'])
+def datatable():
+    
+    Clientes = clientes.query.all()
+    clientes_array = []
+    for cliente in Clientes:
+        cliente_dict = {
+            'nombre': cliente.Nombre,
+            'cedula': cliente.Cedula,
+            'telefono': cliente.Telefono,
+            'direccion': cliente.Direccion,
+        }
+        clientes_array.append(cliente_dict)
+    return jsonify({ 'data': clientes_array })
+
+
+@Clientes.route('/buscar_cliente', methods=['POST'])
+def buscar_cliente_handler():
+    termino_busqueda = request.form['termino_busqueda']
+    Clientes = clientes.query.filter(clientes.Nombre.ilike(f'%{termino_busqueda}%')).all()
+    return render_template('Clientes/Inicio.html', Clientes=Clientes)
 
 
 @Clientes.route('/AgregarCliente', methods = ['GET'])
@@ -18,35 +51,85 @@ def AgregarClienteGet():
 
 @Clientes.route('/AgregarCliente', methods = ['POST'])
 def AgregarClientePost():
-    Nombre= request.form['nombre']
-    Direccion=request.form['direccion']
-    Telefono=request.form['telefono']
-    Cedula=request.form['cedula']
-    
-    new_cliente=clientes(Cedula,Nombre,Direccion,Telefono)
-    db.session.add(new_cliente)
-    db.session.commit()
-    return redirect('/AgregarCliente')
+    try:
+        Nombre = request.form['nombre']
+        Direccion = request.form['direccion']
+        Telefono = request.form['telefono']
+        Cedula = request.form['cedula']
+        
+     
+        if not Nombre or not Direccion or not Telefono or not Cedula:
+            flash("Error: Uno o más campos están vacíos.", "error")
+            return redirect('/AgregarCliente')
+        
+        lista_clientes = clientes.query.all()
+        
+        for cliente in lista_clientes:
+            if cliente.Cedula == Cedula:
+                flash("Error: La cédula ya existe.", "error")
+                return redirect('/AgregarCliente')
+        
+        if not re.match(r'^\d{3}-\d{6}-\d{4}[A-Za-z]$', Cedula):
+            flash("Error: El formato de la cédula no es válido.", "error")
+            return redirect('/AgregarCliente') 
+
+        new_cliente = clientes(Cedula, Nombre, Direccion, Telefono)
+        db.session.add(new_cliente)
+        db.session.commit()
+
+        flash("Cliente agregado correctamente.", "success")
+        return redirect('/AgregarCliente')
+    except KeyError as e:
+        flash(f"Error: {e}. Campo faltante en el formulario.", "error")
+        return redirect('/AgregarCliente')
    
-@Clientes.route('/EliminarCliente/<Cedula>')
+   
+   
+@Clientes.route('/EliminarCliente/<Cedula>', methods=['POST'])
 def EliminarCliente(Cedula):
-    Cliente=clientes.query.get(Cedula)
+    Cliente = clientes.query.get(Cedula)
+    if not Cliente:
+        flash("Error: El cliente no existe.", "error")
+        return redirect('/InicioCliente')
+    
     db.session.delete(Cliente)
     db.session.commit()
-    return redirect("/InicioCliente")
+    return jsonify({"message": "Cliente eliminado correctamente.", "status": "success", "customer": Cliente.serialize()})
     
 
-@Clientes.route('/ActualizarCliente',methods=['POST'])
+@Clientes.route('/ActualizarCliente', methods=['POST'])
 def ActualizarCliente():
-    Cedula=request.form.get('cedula')
+    try:
+        Cedula = request.form.get('cedula')
+        Nombre = request.form['nombre']
+        Direccion = request.form['direccion']
+        Telefono = request.form['telefono']
     
-    Cliente=clientes.query.get(Cedula)
-    Cliente.Nombre=request.form.get('nombre')
-    Cliente.Telefono=request.form.get('telefono')
-    Cliente.Direccion=request.form.get('direccion')
-    db.session.commit()
-    
-    return redirect("/InicioCliente")
+        
+        if  not Nombre or not Direccion or not Telefono:
+            flash("Error: Uno o más campos están vacíos.", "error")
+            return redirect('/InicioCliente')
+        
+        Cliente = clientes.query.get(Cedula)
+        
+        if not Cliente:
+            flash("Error: El cliente no existe.", "error")
+            return redirect('/InicioCliente')
+        
+        Cliente.Nombre = Nombre
+        Cliente.Telefono = Telefono
+        Cliente.Direccion = Direccion
+        db.session.commit()
+        
+        flash("Cliente actualizado correctamente.", "success")
+        return redirect('/InicioCliente')
+        
+    except KeyError as e:
+        flash(f"Error: {e}. Campo faltante en el formulario.", "error")
+        return redirect('/AgregarCliente')
+    except Exception as e:
+        flash(f"Error al actualizar el cliente: {str(e)}", "error")
+        return redirect('/InicioCliente')
 
 
 @Clientes.route('/ObtenerCliente/<Cedula>', methods=['POST'])
